A function to align genotype and phenotype data

A function to align genotype and phenotype data

Usage

add_external_phenotype(
  geno,
  geno_id = "sample.ID",
  pheno,
  pheno_id,
  pheno_col,
  outfile
)

Arguments

geno

An object created by name_and_count_bigsnp()

geno_id

A character string indicating the ID column name in the 'fam' element of the genotype data list. Defaults to 'sample.ID', equivalent to 'IID' in PLINK.

pheno

A data frame with at least two columns: and ID column and a phenotype column

pheno_id

A string specifying the name of the ID column in pheno

pheno_col

A string specifying the name of the phenotype column in pheno. This column will be used as the default y argument to 'plmm()'.

outfile

A string with the name of the filepath for the log file

Value

    Nothing is returned; instead, the results are written to an RDS file.
             By default, the RDS file in `geno` is overwritten. You may change this behavior by specifying a new
             filename to `rdsfile`