A helper function to add predictors to a filebacked matrix of data

A helper function to add predictors to a filebacked matrix of data

Usage

add_predictors_to_bigsnp(
  obj,
  add_predictor_fam,
  add_predictor_ext,
  id_var,
  og_plink_ids,
  quiet
)

Arguments

obj

A bigSNP object

add_predictor_fam

Optional: if you want to include "sex" (the 5th column of .fam file) in the analysis, specify 'sex' here.

add_predictor_ext

Optional: add additional covariates/predictors/features from an external file (i.e., not a PLINK file).

id_var

String specifying which column of the PLINK .fam file has the unique sample identifiers. Options are "IID" (default) and "FID".

og_plink_ids

Character vector passed from name_and_count_bigsnp()

quiet

Logical: should messages be printed to the console? Defaults to TRUE

Value

A list of 2 components:

'obj' - a bigSNP object with an added element representing the matrix that includes the additional predictors as the first few columns

'non_gen' - an integer vector that ranges from 1 to the number of added predictors. Example: if 2 predictors are added, non_gen = 1:2