Coef method for "cv_plmm" class

Coef method for "cv_plmm" class

Usage

# S3 method for class 'cv_plmm'
coef(object, lambda, which = object$min, ...)

Arguments

object

An object of class "cv_plmm."

lambda

A numeric vector of lambda values.

which

Vector of lambda indices for which coefficients to return. Defaults to lambda index with minimum CVE.

...

Additional arguments (not used).

Value

Returns a named numeric vector. Values are the coefficients of the model at the specified value of either lambda or which. Names are the values of lambda.

Examples

cv_fit <- cv_plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
head(coef(cv_fit))
#> (Intercept)        Var1        Var2        Var3        Var4        Var5 
#>  5.69945122 -0.22484261  0.00000000  3.24422663 -0.02824771  0.00000000