Cross-validation for plmm

Performs k-fold cross validation for lasso-, MCP-, or SCAD-penalized linear mixed models over a grid of values for the regularization parameter lambda.

Usage

cv.plmm(
  X,
  y,
  k = NULL,
  K = NULL,
  diag_K = NULL,
  eta_star = NULL,
  penalty = c("MCP", "SCAD", "lasso"),
  penalty.factor = rep(1, ncol(X)),
  type = "blup",
  cluster,
  nfolds = 10,
  seed,
  fold = NULL,
  returnY = FALSE,
  returnBiasDetails = FALSE,
  trace = FALSE,
  ...
)

Arguments

X

Design matrix for model fitting. May include clinical covariates and other non-SNP data.

y

Continuous outcome vector for model fitting.

k

An integer specifying the number of singular values to be used in the approximation of the rotated design matrix. This argument is passed to RSpectra::eigs(). Defaults to min(n, p) - 1, where n and p are the dimensions of the standardized design matrix.

K

Similarity matrix, in the form of (1) the relatedness matrix estimated from the data (default), (2) a user-supplied matrix, or (3) a user-supplied list with components 'd' and 'u' as created by choose_k().

diag_K

Logical: should K be a diagonal matrix? This would reflect observations that are unrelated, or that can be treated as unrelated. Defaults to FALSE.

eta_star

Optional arg. to plmm_prep. Defaults to NULL.

penalty

The penalty to be applied to the model. Either "MCP" (the default), "SCAD", or "lasso".

penalty.factor

Optional arg. to plmm_prep. Defaults to 1 for all predictors (except the intercept).

type

A character argument indicating what should be returned from predict.plmm(). If type == 'lp' predictions are based on the linear predictor, $X beta$. If type == 'blup' predictions are based on the sum of the linear predictor and the estimated random effect (BLUP). Defaults to 'blup', as this has shown to be a superior prediction method in many applications.

cluster

cv.plmm() can be run in parallel across a cluster using the parallel package. The cluster must be set up in advance using parallel::makeCluster(). The cluster must then be passed to cv.plmm().

nfolds

The number of cross-validation folds. Default is 10.

seed

You may set the seed of the random number generator in order to obtain reproducible results.

fold

Which fold each observation belongs to. By default the observations are randomly assigned.

returnY

Should cv.plmm() return the linear predictors from the cross-validation folds? Default is FALSE; if TRUE, this will return a matrix in which the element for row i, column j is the fitted value for observation i from the fold in which observation i was excluded from the fit, at the jth value of lambda.

returnBiasDetails

Logical: should the cross-validation bias (numeric value) and loss (n x p matrix) be returned? Defaults to FALSE.

trace

If set to TRUE, inform the user of progress by announcing the beginning of each CV fold. Default is FALSE.

...

Additional arguments to plmm_fit

Value

a list with 11 items:

type: the type of prediction used ('lp' or 'blup')

cve: numeric vector with the cross validation error (CVE) at each value of lambda

cvse: numeric vector with the estimated standard error associated with each value of for cve

fold: numeric n length vector of integers indicating the fold to which each observation was assigned

lambda: numeric vector of lambda values

fit: the overall fit of the object, including all predictors; this is a list as returned by plmm()

min: The index corresponding to the value of lambda that minimizes cve

lambda.min: The lambda value at which cve is minmized

min1se: The index corresponding to the value of lambda within standard error of that which minimizes cve

lambda1se: largest value of lambda such that error is within 1 standard error of the minimum.

null.dev: numeric value representing the deviance for the intercept-only model. If you have supplied your own lambda sequence, this quantity may not be meaningful.

Examples

cv_fit <- cv.plmm(X = admix$X, y = admix$y, seed = 321)
if (FALSE) {
cv_s <- summary.cv.plmm(cv_fit, lambda = "1se")
print(cv_s)
plot(cv_fit)
}