A function to convert the output of process_plink() into a big.matrix (such as could be passed to biglasso::biglasso())

A function to convert the output of process_plink() into a big.matrix (such as could be passed to biglasso::biglasso())

Usage

get_bm_from_plink(path, standardize = TRUE)

Arguments

path

The file path to the RDS object containing the processed data. Do not add the '.rds' extension to the path.

standardize

Logical: should the data be returned in a column-standardized form? Defaults to TRUE. Please don't change this unless you are super confident about what you are doing...

Value

A list with elements:

'X': a big.matrix object with the design matrix of the data. If standardize = TRUE, the data corresponding to std_X is returned. Otherwise, the data corresponding to subset_X is returned

'map': data frame with .bim file information

'fam': data frame with .fam file information

'ns': A vector indicating the which columns of X contain nonsingular features (i.e., features with variance != 0.

'center': If standardize = TRUE, a vector of values for centering each column in X is returned

'scale': If standardize = TRUE, a vector of values for scaling each column in X is returned

Details

This function is a wrapper combining get_data() and fbm2bm()

Examples

# \donttest{
process_plink(data_dir = get_example_data(parent = TRUE),
  prefix = "penncath_lite",
  gz = TRUE,
  outfile = "process_penncath",
  # overwrite = TRUE, # uncomment if needed 
  impute_method = "mode")
#> 
#> Logging to process_penncath.log
#> Preprocessing penncath_lite data:
#> Creating  penncath_lite .rds
#> 
#> Unzipping .gz files - this could take a second
#> There are  1401  observations and  4367  genomic features in the specified data files, representing chromosomes  1  -  22
#> There are a total of  3514 SNPs with missing values
#> Of these,  13  are missing in at least 50% of the samples
#> Will prune out  468  samples/observations with missing phenotype data.
#> Imputing the missing (genotype) values using  mode  method
#> Error: Two levels of parallelism are used. See `?assert_cores`.
  
  my_path <- paste0(get_example_data(parent = TRUE), "/penncath_lite")
  bm_data <- get_bm_from_plink(my_path)
#> Note: The design matrix is being returned as a file-backed matrix (FBM) -- see bigstatsr::FBM() for details.
#> Reminder: the X that is returned here is column-standardized.
#>         
#> A copy of the original data is available via the 'genotypes' matrix in the .rds object
#> Error in if (obj$id_var == "IID") {    row_names <- as.character(obj$fam$sample.ID[obj$complete_phen])}: argument is of length zero
# }