A helper function to handle missing genotype values from PLINK data

A helper function to handle missing genotype values from PLINK data

Usage

handle_missingness(
  obj,
  X,
  counts,
  na_phenotype_vals,
  handle_missing_phen,
  outfile,
  quiet
)

Arguments

obj

a bigSNP object

X

A matrix of genotype data as returned by name_and_count_bigsnp

counts

A dataframe as created by bigstatsr::big_counts() and returned by name_and_count_bigsnp

na_phenotype_vals

A vector of numeric values used to code NA values in the phenotype/outcome (this is the 'affection' column in a bigSNP object, or the last column of a .fam file). Defaults to -9 (matching PLINK conventions).

handle_missing_phen

A string indicating how missing phenotypes should be handled:

"prune" (default): observations with missing phenotype are removed

"asis": leaves missing phenotypes as NA (this is fine if outcome will be supplied later from a separate file)

"median": impute missing phenotypes using the median (warning: this is overly simplistic in many cases).

"mean": impute missing phenotypes using the mean (warning: this is overly simplistic in many cases).

outfile

Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.

quiet

Logical: should messages be printed to the console? Defaults to TRUE

Value

a list of two components:

na_idx: a logical vector indicating which columns have missing values

prop_na: the proportion of missingness in each of the columns tagged by 'na_idx'