plmmr: Penalized Linear Mixed Models For Correlated Data

Fits penalized linear mixed models that correct for unobserved confounding factors. 'plmmr' infers and corrects for the presence of unobserved confounding effects such as population stratification and environmental heterogeneity. It then fits a linear model via penalized maximum likelihood. Originally designed for the multivariate analysis of single nucleotide polymorphisms (SNPs) measured in a genome-wide association study (GWAS), 'plmmr' eliminates the need for subpopulation-specific analyses and post-analysis p-value adjustments. Functions for the appropriate processing of 'PLINK' files are also supplied. For examples, see the package homepage. https://pbreheny.github.io/plmmr/.

Author

Maintainer: Patrick J. Breheny patrick-breheny@uiowa.edu (ORCID)

Authors:

Anna C. Reisetter anna-reisetter@uiowa.edu (ORCID)

Tabitha K. Peter tabitha-peter@uiowa.edu (ORCID)

Yujing Lu