A helper function to standardize a bigSNP

A helper function to standardize a bigSNP

Usage

standardize_bigsnp(
  obj,
  prefix,
  rds_dir,
  non_gen,
  complete_phen,
  id_var,
  outfile,
  quiet,
  overwrite
)

Arguments

obj

A bigSNP object

prefix

The prefix (as a character string) of the bed/fam data files (e.g., prefix = 'mydata')

rds_dir

The path to the directory in which you want to create the new '.rds' and '.bk' files. Defaults to data_dir

non_gen

An integer vector that ranges from 1 to the number of added predictors. Example: if 2 predictors are added, non_gen = 1:2. Note: this is typically passed from the result of add_predictors()

complete_phen

Numeric vector with indicies marking the rows of the original data which have a non-missing entry in the 6th column of the .fam file

id_var

String specifying which column of the PLINK .fam file has the unique sample identifiers. Options are "IID" (default) and "FID".

outfile

Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.

quiet

Logical: should messages be printed to the console? Defaults to TRUE

overwrite

Logical: if existing .bk/.rds files exist for the specified directory/prefix, should these be overwritten?

Value

A list with a new component of obj called 'std_X' - this is an FBM with column-standardized data. List also includes several other indices/meta-data on the standardized matrix