A summary function for cv.plmm objects

A summary function for cv.plmm objects

Usage

# S3 method for cv.plmm
summary(object, lambda = "min", ...)

Arguments

object

A cv.plmm object

lambda

The regularization parameter value at which inference should be reported. Can choose a numeric value, 'min', or '1se'. Defaults to 'min.'

...

Not used

Value

The return value is an object with S3 class summary.cv.plmm. The class has its own print method and contains the following list elements:

lambda.min: The lambda value at the minimum cross validation error

lambda.1se: The maximum lambda value within 1 standard error of the minimum cross validation error

nvars: The number of non-zero coefficients at the selected lambda value

cve: The cross validation error at all folds

min: The minimum cross validation error

fit: The plmm fit used in the cross validation

bias: The mean bias of the cross validation

loss: The loss (at each fold?) TODO: double-check this

penalty: The penalty applied to the fitted model

Examples

cv_fit <- cv.plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
summary(cv_fit)
#> MCP-penalized model with n=197 and p=100
#> At minimum cross-validation error (lambda=0.4279):
#> -------------------------------------------------
#>   Nonzero coefficients: 0
#>   Cross-validation error (deviance): 2.07
#>   Scale estimate (sigma): 1.440