Coef method for "plmm" class

Coef method for "plmm" class

Usage

# S3 method for class 'plmm'
coef(object, lambda, which = 1:length(object$lambda), drop = TRUE, ...)

Arguments

object

An object of class "plmm."

lambda

A numeric vector of lambda values.

which

Vector of lambda indices for which coefficients to return.

drop

Logical.

...

Additional arguments.

Value

Either a numeric matrix (if model was fit on data stored in memory) or a sparse matrix (if model was fit on data stored filebacked). Rownames are feature names, columns are values of lambda.

Examples

fit <- plmm(admix$X, admix$y)
coef(fit)[1:10, 1:5]
#>              0.42141  0.39301  0.36652  0.34182 0.31878
#> (Intercept) 4.145401 4.156508 4.162843 4.124933  4.1066
#> Snp1        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp2        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp3        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp4        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp5        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp6        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp7        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp8        0.000000 0.000000 0.000000 0.000000  0.0000
#> Snp9        0.000000 0.000000 0.000000 0.000000  0.0000