Read in processed data This function is intended to be called after either process_plink() or process_delim() has been called once.

Read in processed data This function is intended to be called after either process_plink() or process_delim() has been called once.

Usage

get_data(path, returnX, trace = TRUE)

Arguments

path

The file path to the RDS object containing the processed data. Do not add the '.rds' extension to the path.

returnX

Logical: Should the design matrix be returned as a numeric matrix that will be stored in memory. By default, this will be FALSE if the object sizes exceeds 100 Mb.

trace

Logical: Should trace messages be shown? Default is TRUE.

Value

A list with these components:

std_X, the column-standardized design matrix as either (1) a numeric matrix or (2) a filebacked matrix (FBM). See bigstatsr::FBM() and bigsnpr::bigSnp-class documentation for details.

fam, a data frame containing the pedigree information (like a .fam file in PLINK)

map, a data frame containing the feature information (like a .bim file in PLINK)

ns: A vector indicating the which columns of X contain nonsingular features (i.e., features with variance != 0.

center: A vector of values for centering each column in X

scale: A vector of values for scaling each column in X

Details

The .rds object should have an 'std_X' element - this is what will be used as the design matrix for analysis. This design matrix should not include an intercept column (this will be added later in plmm_fit()).

In the returned list, the fam data will be sorted by family and by individual, as in dplyr::arrange(family.ID, sample.ID). The rows of X will be sorted to align in the same order as in fam, where rownames of X will be sample ID.

Examples

# \donttest{
temp_dir <- paste0(tempdir()) # using a temporary directory here
process_plink(data_dir = find_example_data(parent = TRUE), # reads data that ships with plmmr
              rds_dir = temp_dir,
              prefix = "penncath_lite",
              gz = TRUE,
              outfile = "process_penncath",
              overwrite = TRUE,
              impute_method = "mode")
#> 
#> Logging to process_penncath.log
#> Preprocessing penncath_lite data:
#> It looks like your files are zipped -- please unzip them before calling process_plink().
#> Creating penncath_lite.rds
#> Error: File '/home/runner/work/_temp/Library/plmmr/extdata/penncath_lite.bed' doesn't exist.

 pen <- get_data(file.path(temp_dir, "std_penncath_lite"))
#> Warning: cannot open compressed file '/tmp/RtmpoE4e8y/std_penncath_lite.rds', probable reason 'No such file or directory'
#> Error in gzfile(file, "rb"): cannot open the connection
 str(pen)
#> Error in eval(expr, envir, enclos): object 'pen' not found
# }