plmm_checks

plmm_checks

Usage

plmm_checks(
  X,
  y = NULL,
  std_needed = NULL,
  col_names = NULL,
  non_genomic = NULL,
  K = NULL,
  diag_K = NULL,
  eta_star = NULL,
  penalty = c("MCP", "SCAD", "lasso"),
  penalty.factor = NULL,
  init = NULL,
  gamma,
  alpha = 1,
  dfmax = NULL,
  trace = FALSE,
  ...
)

Arguments

X

Design matrix object or a string with the file path to a design matrix. If a string, string will be passed to get_data().

Note: X may include clinical covariates and other non-SNP data, but no missing values are allowed.

y

Continuous outcome vector. Defaults to NULL, assuming that the outcome is the 6th column in the .fam PLINK file data. Can also be a user-supplied numeric vector.

std_needed

Logical: does the supplied X need to be standardized? Defaults to NULL, since by default, X will be standardized internally. For data processed from PLINK files, standardization happens in process_plink(). For data supplied as a matrix, standardization happens here in plmm(). If you know your data are already standardized, leave std_needed = FALSE -- this would be an atypical case. Note: failing to standardize data will lead to incorrect analyses.

col_names

Optional vector of column names for design matrix. Defaults to NULL.

non_genomic

Optional vector specifying which columns of the design matrix represent features that are not genomic, as these features are excluded from the empirical estimation of genomic relatedness. For cases where X is a filepath to an object created by process_plink(), this is handled automatically via the arguments to process_plink(). For all other cases, 'non_genomic' defaults to NULL (meaning plmm() will assume that all columns of X represent genomic features).

K

Similarity matrix used to rotate the data. This should either be (1) a known matrix that reflects the covariance of y, (2) an estimate (Default is \(\frac{1}{p}(XX^T)\)), or (3) a list with components 'd' and 'u', as returned by choose_k().

diag_K

Logical: should K be a diagonal matrix? This would reflect observations that are unrelated, or that can be treated as unrelated. Defaults to FALSE. Note: plmm() does not check to see if a matrix is diagonal. If you want to use a diagonal K matrix, you must set diag_K = TRUE.

eta_star

Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.

penalty

The penalty to be applied to the model. Either "MCP" (the default), "SCAD", or "lasso".

penalty.factor

A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to the number of columns of X. The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. In particular, penalty.factor can be 0, in which case the coefficient is always in the model without shrinkage.

init

Initial values for coefficients. Default is 0 for all columns of X.

gamma

The tuning parameter of the MCP/SCAD penalty (see details). Default is 3 for MCP and 3.7 for SCAD.

alpha

Tuning parameter for the Mnet estimator which controls the relative contributions from the MCP/SCAD penalty and the ridge, or L2 penalty. alpha=1 is equivalent to MCP/SCAD penalty, while alpha=0 would be equivalent to ridge regression. However, alpha=0 is not supported; alpha may be arbitrarily small, but not exactly 0.

dfmax

Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.

trace

If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.

...

Additional arguments to get_data()