PLMM fit: a function that fits a PLMM using the values returned by plmm_prep() This is an internal function for cv.plmm

PLMM fit: a function that fits a PLMM using the values returned by plmm_prep() This is an internal function for cv.plmm

Usage

plmm_fit(
  prep,
  penalty = "MCP",
  gamma,
  alpha = 1,
  lambda.min,
  nlambda = 100,
  lambda,
  eps = 1e-04,
  max.iter = 10000,
  convex = TRUE,
  dfmax = prep$p + 1,
  init = NULL,
  warn = TRUE,
  returnX = TRUE
)

Arguments

prep

A list as returned from plmm_prep

penalty

The penalty to be applied to the model. Either "MCP" (the default), "SCAD", or "lasso".

gamma

The tuning parameter of the MCP/SCAD penalty (see details). Default is 3 for MCP and 3.7 for SCAD.

alpha

Tuning parameter for the Mnet estimator which controls the relative contributions from the MCP/SCAD penalty and the ridge, or L2 penalty. alpha=1 is equivalent to MCP/SCAD penalty, while alpha=0 would be equivalent to ridge regression. However, alpha=0 is not supported; alpha may be arbitrarily small, but not exactly 0.

lambda.min

The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .05 otherwise.

nlambda

Length of the sequence of lambda. Default is 100.

lambda

A user-specified sequence of lambda values. By default, a sequence of values of length nlambda is computed, equally spaced on the log scale.

eps

Convergence threshold. The algorithm iterates until the RMSD for the change in linear predictors for each coefficient is less than eps. Default is 1e-4.

max.iter

Maximum number of iterations (total across entire path). Default is 10000.

convex

convex Calculate index for which objective function ceases to be locally convex? Default is TRUE.

dfmax

(future idea; not yet incorporated) Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.

init

Initial values for coefficients. Default is 0 for all columns of X.

warn

Return warning messages for failures to converge and model saturation? Default is TRUE.

returnX

Return the standardized design matrix along with the fit? By default, this option is turned on if X is under 100 MB, but turned off for larger matrices to preserve memory.

Value

A list with these components:

std_X: The standardized design matrix

rot_X: first partial result of data rotation

rot_y: second partial result of data rotation

eta: numeric value representing the ratio of variances.

std_rot_X: re-standardized rotated design matrix. This is 'fed' into plmm_fit().

b: The values returned in the 'beta' argument of the ncvfit() object

lambda: The sequence of lambda values used in model fitting

iter: The number of iterations at each given lambda value

converged: The convergence status at each given lambda value

penalty: The type of penalty used in model fitting

penalty.factor: A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to the number of columns of X. The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. In particular, penalty.factor can be 0, in which case the coefficient is always in the model without shrinkage.

ns: The indices of the non-singular columns of the ORIGINAL design matrix

ncol_X: The number of columns in the ORIGINAL design matrix