PLMM fit: a function that fits a PLMM using the values returned by plmm_prep() This is an internal function for cv.plmm

PLMM fit: a function that fits a PLMM using the values returned by plmm_prep() This is an internal function for cv.plmm

Usage

plmm_fit(
  prep,
  penalty = "MCP",
  gamma,
  alpha = 1,
  lambda.min,
  nlambda = 100,
  lambda,
  eps = 1e-04,
  max.iter = 10000,
  convex = TRUE,
  dfmax = prep$p + 1,
  init = NULL,
  warn = TRUE,
  returnX = TRUE
)

Arguments

prep

A list as returned from plmm_prep

penalty

The penalty to be applied to the model. Either "MCP" (the default), "SCAD", or "lasso".

gamma

The tuning parameter of the MCP/SCAD penalty (see details). Default is 3 for MCP and 3.7 for SCAD.

alpha

Tuning parameter for the Mnet estimator which controls the relative contributions from the MCP/SCAD penalty and the ridge, or L2 penalty. alpha=1 is equivalent to MCP/SCAD penalty, while alpha=0 would be equivalent to ridge regression. However, alpha=0 is not supported; alpha may be arbitrarily small, but not exactly 0.

lambda.min

The smallest value for lambda, as a fraction of lambda.max. Default is .001 if the number of observations is larger than the number of covariates and .05 otherwise.

nlambda

Length of the sequence of lambda. Default is 100.

lambda

A user-specified sequence of lambda values. By default, a sequence of values of length nlambda is computed, equally spaced on the log scale.

eps

Convergence threshold. The algorithm iterates until the RMSD for the change in linear predictors for each coefficient is less than eps. Default is 1e-4.

max.iter

Maximum number of iterations (total across entire path). Default is 10000.

convex

convex Calculate index for which objective function ceases to be locally convex? Default is TRUE.

dfmax

(future idea; not yet incorporated) Upper bound for the number of nonzero coefficients. Default is no upper bound. However, for large data sets, computational burden may be heavy for models with a large number of nonzero coefficients.

init

Initial values for coefficients. Default is 0 for all columns of X.

warn

Return warning messages for failures to converge and model saturation? Default is TRUE.

returnX

Return the standardized design matrix along with the fit? By default, this option is turned on if X is under 100 MB, but turned off for larger matrices to preserve memory.

Value

A list with these components:

n: # of rows in X

p: # of columns in X (including constant features)

y: outcome on original scale

std_X_details: list with 'center' and 'scale' vectors, same as plmm_prep()

s: eigenvalues of K

U: eigenvectors of K

rot_X: X on the rotated (i.e., transformed) scale. Note that the dimensions of rot_X are likely to be different than those of X.

rot_y: y on the rotated scale

stdrot_X: X on the rotated scale once it has been re-standardized.

lambda: vector of tuning parameter values

b: the coefficients estimated on the scale of stdrot_X

untransformed_b1: the coefficients estimated on the scale of std_X

linear.predictors: the product of stdrot_X and b (linear predictors on the transformed and restandardized scale)

eta: a number (double) between 0 and 1 representing the estimated proportion of the variance in the outcome attributable to population/correlation structure.

iter: numeric vector with the number of iterations needed in model fitting for each value of lambda

converged: vector of logical values indicating whether the model fitting converged at each value of lambda

loss: vector with the numeric values of the loss at each value of lambda (calculated on the ~rotated~ scale)

penalty: character string indicating the penalty with which the model was fit (e.g., 'MCP')

penalty.factor: vector of indicators corresponding to each predictor, where 1 = predictor was penalized.

gamma: numeric value indicating the tuning parameter used for the SCAD or lasso penalties was used. Not relevant for lasso models.

alpha: numeric value indicating the elastic net tuning parameter.

ns: the indices for the nonsingular values of X

snp_names: ormatted column names of the design matrix

nlambda: number of lambda values used in model fitting

eps: tolerance ('epsilon') used for model fitting

max.iter: max. number of iterations per model fit

warn: logical - should warnings be given if model fit does not converge?

init: initial values for model fitting

trace: logical - should messages be printed to the console while models are fit?