PLMM prep: a function to run checks, SVD, and rotation prior to fitting a PLMM model This is an internal function for cv.plmm

PLMM prep: a function to run checks, SVD, and rotation prior to fitting a PLMM model This is an internal function for cv.plmm

Usage

plmm_prep(
  X,
  y,
  k = NULL,
  K = NULL,
  diag_K = NULL,
  eta_star = NULL,
  penalty.factor = rep(1, ncol(X)),
  trace = NULL,
  ...
)

Arguments

X

Design matrix. May include clinical covariates and other non-SNP data.

y

Continuous outcome vector.

k

An integer specifying the number of singular values to be used in the approximation of the rotated design matrix. This argument is passed to RSpectra::svds(). Defaults to min(n, p) - 1, where n and p are the dimensions of the standardized design matrix.

K

Similarity matrix used to rotate the data. This should either be a known matrix that reflects the covariance of y, or an estimate (Default is \(\frac{1}{p}(XX^T)\), where X is standardized). This can also be a list, with components d and u (as returned by choose_k)

diag_K

Logical: should K be a diagonal matrix? This would reflect observations that are unrelated, or that can be treated as unrelated. Passed from plmm().

eta_star

Optional argument to input a specific eta term rather than estimate it from the data. If K is a known covariance matrix that is full rank, this should be 1.

penalty.factor

A multiplicative factor for the penalty applied to each coefficient. If supplied, penalty.factor must be a numeric vector of length equal to the number of columns of X. The purpose of penalty.factor is to apply differential penalization if some coefficients are thought to be more likely than others to be in the model. In particular, penalty.factor can be 0, in which case the coefficient is always in the model without shrinkage.

trace

If set to TRUE, inform the user of progress by announcing the beginning of each step of the modeling process. Default is FALSE.

...

Not used yet

Value

List with these components:

n: the number of rows in the original design matrix

p: the number of columns in the original design matrix

y: The vector of outcomes

std_X: standardized design matrix

std_X_details: a list with 2 vectors:

'center' (values used to center X)

'scale' (values used to scale X)

s: vector with the eigenvalues of K

U: the eigenvectors of K (same as left singular values of X).

ns: the indices for the nonsingular values of X

penalty.factor: the penalty factors for the penalized non-singular values

snp_names: formatted column names of the design matrix

Examples

if (FALSE) {
# this is an internal function; to call this, you would need to use the triple 
# colon, e.g., plmmr:::plmm_prep()
prep1 <- plmm_prep(X = admix$X, y = admix$y, trace = TRUE)
prep2 <- plmm_prep(X = admix$X, y = admix$y, diag_K = TRUE, trace = TRUE)
}