A function to read in large data files as an FBM

A function to read in large data files as an FBM

Usage

process_delim(
  file,
  data_dir,
  rds_dir = data_dir,
  ind.col,
  non_gen = NULL,
  outfile,
  overwrite = FALSE,
  quiet = FALSE
)

Arguments

file

The file to be read in, without the filepath. This should be a file of numeric values, no header row! Headers should be taken out of the file and supplied to the col_names argument in plmm() Example: use file = "myfile.txt", not file = "~/mydirectory/myfile.txt"

data_dir

The directory to the file

rds_dir

The directory where the user wants to create the '.rds' and '.bk' files Defaults to data_dir

ind.col

Numeric vector of the columns to read in. Don't use negative indicies. If you're not sure how many columns are in your file, use something like data.table::fread() to examine the first row or two.

non_gen

Numeric vector with indices of columns that have non-genomic information. Note: if you plan on estimating the genomic relatedness among observations in your file, you must include this information in order for the estimates to be correct.

outfile

Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_delim', i.e. you will get 'process_delim.log' as the outfile.

overwrite

Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.

quiet

Logical: should the messages printed to the console be silenced? Defaults to FALSE.

Value

Nothing is returned by this function, but (at least) two files are created in the location specified by rds_dir:

'std_prefix.rds': This is the bigsnpr::bigSNP object that holds the PLINK data along with meta-data. See details for explanation of what is included in this meta-data

'std_prefix.bk': Created by the call to standardize_fbm(), this is the backingfile that stores the numeric data of the standardized design matrix std_X

Examples

# \donttest{
temp_dir <- tempdir()
process_delim(file = "colon2.txt",
 data_dir = get_example_data(parent = TRUE),
 rds_dir = temp_dir,
  ind.col = 2:2002)
#> 
#> Logging to /tmp/RtmpJe91X5/process_data.log
#> Preprocessing colon2 data:
#> There are a total of  20 features with missing values. 
#>       
#> At this time, plmmr::process_delim() does not impute missing values. We 
#>       are working to develop this feature. 
#>       
#> For now, plmmr::process_delim() will drop all columns of X with NA values.
#> Column-standardizing the design matrix...
#> Error: Two levels of parallelism are used. See `?assert_cores`.

# }