Preprocess PLINK files using the bigsnpr package

Preprocess PLINK files using the bigsnpr package

Usage

process_plink(
  data_dir,
  prefix,
  rds_dir = data_dir,
  impute = TRUE,
  impute_method = "mode",
  na_phenotype_vals = c(-9),
  id_var = "IID",
  handle_missing_phen = "prune",
  add_phen = NULL,
  pheno_id = NULL,
  pheno_name = NULL,
  quiet = FALSE,
  gz = FALSE,
  outfile,
  overwrite = FALSE,
  add_predictor_fam = NULL,
  add_predictor_ext = NULL,
  keep_bigSNP = FALSE,
  ...
)

Arguments

data_dir

The path to the bed/bim/fam data files, without a trailing "/" (e.g., use data_dir = '~/my_dir', not data_dir = '~/my_dir/')

prefix

The prefix (as a character string) of the bed/fam data files (e.g., prefix = 'mydata')

rds_dir

The path to the directory in which you want to create the new '.rds' and '.bk' files. Defaults to data_dir

impute

Logical: should data be imputed? Default to TRUE.

impute_method

If 'impute' = TRUE, this argument will specify the kind of imputation desired. Options are: * mode (default): Imputes the most frequent call. See bigsnpr::snp_fastImputeSimple() for details. * random: Imputes sampling according to allele frequencies. * mean0: Imputes the rounded mean. * mean2: Imputes the mean rounded to 2 decimal places. * xgboost: Imputes using an algorithm based on local XGBoost models. See bigsnpr::snp_fastImpute() for details. Note: this can take several minutes, even for a relatively small data set.

na_phenotype_vals

A vector of numeric values used to code NA values in the phenotype/outcome (this is the 'affection' column in a bigSNP object, or the last column of a .fam file). Defaults to -9 (matching PLINK conventions).

id_var

String specifying which column of the PLINK .fam file has the unique sample identifiers. Options are "IID" (default) and "FID".

handle_missing_phen

A string indicating how missing phenotypes should be handled: * "prune" (default): observations with missing phenotype are removed * "median": impute missing phenotypes using the median (warning: this is overly simplistic in many cases). * "mean": impute missing phenotypes using the mean (warning: this is overly simplistic in many cases). Note: for data coming from PLINK, no missing values of the phenotype are allowed. You have to (1) supply phenotype from an external file, (2) prune missing values, or (3) impute missing values.

add_phen

Optional: A data frame with at least two columns: and ID column and a phenotype column

pheno_id

Optional: A string specifying the name of the ID column in pheno. MUST be specified if add_phen is specified.

pheno_name

Optional: A string specifying the name of the phenotype column in pheno. MUST be specified if add_phen is specified. This column will be used as the default y argument to 'plmm()'.

quiet

Logical: should messages to be printed to the console be silenced? Defaults to FALSE

gz

Logical: are the bed/bim/fam files g-zipped? Defaults to FALSE. NOTE: if TRUE, process_plink will unzip your zipped files.

outfile

Optional: the name (character string) of the prefix of the logfile to be written. Defaults to 'process_plink', i.e. you will get 'process_plink.log' as the outfile.

overwrite

Logical: if existing .bk/.rds files exist for the specified directory/prefix, should these be overwritten? Defaults to FALSE. Set to TRUE if you want to change the imputation method you're using, etc.

add_predictor_fam

Optional: if you want to include "sex" (the 5th column of .fam file) in the analysis, specify 'sex' here.

add_predictor_ext

Optional: add additional covariates/predictors/features from an external file (i.e., not a PLINK file). This argument takes one of two kinds of arguments: - a named numeric vector, where the names align with the sample IDs in the PLINK files. The names will be used to subset and align this external covariate with the supplied PLINK data.

                         - a numeric matrix whose row names align with the sample IDs in the PLINK files. 
                      The names will be used to subset and align this external covariate with the supplied PLINK data.

keep_bigSNP

Logical: should the intermediate steps of data processing be saved as a bigSNP object? Defaults to FALSE.

...

Optional: additional arguments to bigsnpr::snp_fastImpute() (relevant only if impute_method = "xgboost")

Value

Nothing is returned by this function, but (at least) two files are created in the location specified by rds_dir:

'std_prefix.rds': This is the bigsnpr::bigSNP object that holds the PLINK data along with meta-data. See details for explanation of what is included in this meta-data

'std_prefix.bk': Created by the call to standardize_fbm(), this is the backingfile that stores the numeric data of the standardized design matrix std_X

Intermediate files 'prefix.rds' and 'prefix.bk' are also created along the way; if keep_bigSNP = TRUE, these are not deleted at the end of the process_plink()

procedure. Note that these files could potentially be quite large - that's the main reason that keeping these files is not the default setting.

Note that process_plink() need only be run once for a given set of PLINK files; in subsequent data analysis/scripts, get_data() will access the '.rds' file.

Details

The '.rds' object created by this function has the following elements:

std_X: The file-backed design matrix, as an FBM object (see bigstatsr package for more info)

fam: Data frame equivalent of PLINK '.bim' file

map: Data frame equivalent of PLINK '.fam' file

colnames: Character vector of column names for the original data (includes constant features, i.e. monomorphic SNPs)

rownames: Character vector of row names for the original data.

n: Number of rows (samples) in the original data

p: Number of columns (features, SNPs, markers, ...) in the original data

ns: Numeric vector of indices marking the non-singular columns of the original data

std_X_center: Numeric vector of values used to center the non-singular columns of the data

std_X_scale: Numeric vector of values used to scale the non-singular columns of the data

std_X_colnames: Character vector of column names for the standardized data

std_X_rownames: Character vector of row names for the standardized data.

complete_phen: Numeric vector of indices marking the samples with a non-missing phenotype. Ony applicable if handle_missing_phen = 'prune'

id_var: String specifying which ID column in the '.fam' file had the unique sample ID: 'FID' (1st column) or 'IID' (2nd column)

Examples

# \donttest{
my_dir <- tempdir()
process_plink(data_dir = get_example_data(parent = TRUE),
  rds_dir = my_dir,
  prefix = "penncath_lite",
  gz = TRUE,
  outfile = "process_penncath",
  overwrite = TRUE,
  impute_method = "mode")
#> 
#> Logging to process_penncath.log
#> Preprocessing penncath_lite data:
#> Overwriting existing files:  penncath_lite .bk/.rds
#> 
#> Creating  penncath_lite .rds
#> 
#> The file  /home/runner/work/_temp/Library/plmmr/extdata/penncath_lite.bed  already exists. Setting gz = FALSE
#> There are  1401  observations and  4367  genomic features in the specified data files, representing chromosomes  1  -  22
#> There are a total of  3514 SNPs with missing values
#> Of these,  13  are missing in at least 50% of the samples
#> Will prune out  468  samples/observations with missing phenotype data.
#> Imputing the missing (genotype) values using  mode  method
#> Error: Two levels of parallelism are used. See `?assert_cores`.
# }