A summary method for the plmm objects

A summary method for the plmm objects

Usage

# S3 method for class 'plmm'
summary(object, lambda, idx, eps = 1e-05, ...)

Arguments

object

An object of class plmm

lambda

The regularization parameter value at which inference should be reported.

idx

Alternatively, lambda may be specified by an index; idx=10 means: report inference for the 10th value of lambda along the regularization path. If both lambda and idx are specified, lambda takes precedence.

eps

If lambda is given, eps is the tolerance for difference between the given lambda value and a lambda value from the object. Defaults to 0.0001 (1e-5)

...

Not used

Value

The return value is an object with S3 class summary.plmm. The class has its own print method and contains the following list elements:

penalty: The penalty used by plmm (e.g. SCAD, MCP, lasso)

n: Number of instances/observations

p: Number of regression coefficients (not including the intercept)

converged: Logical indicator for whether the model converged

lambda: The lambda value at which inference is being reported

lambda_char: A formatted character string indicating the lambda value

nvars: The number of nonzero coefficients (again, not including the intercept) at that value of lambda

nonzero: The column names indicating the nonzero coefficients in the model at the specified value of lambda

constant_features: A character vector with the names of any columns in the design matrix whose values are constant in the whole sample (e.g., monomorphic SNPs in a genetics context)

Examples

fit <- cv_plmm(X = admix$X, y = admix$y, K = relatedness_mat(admix$X))
summary(fit$fit, idx = 97)
#> lasso-penalized regression model with n=197, p=101 at lambda=0.0020
#> -------------------------------------------------
#> The model converged 
#> -------------------------------------------------
#> # of non-zero coefficients:  97 
#> -------------------------------------------------
summary(fit, lambda = fit$lambda_min)
#> lasso-penalized model with n=197 and p=101
#> At minimum cross-validation error (lambda=0.0655):
#> -------------------------------------------------
#>   Nonzero coefficients: 46
#>   Cross-validation error (deviance): 2.25
#>   Scale estimate (sigma): 1.501